+++ packages/framework/src/Form/Admin/Category/CategoryFormType.php
@@ -7,6 +7,7 @@ use Shopsys\FormTypesBundle\MultidomainType;
 use Shopsys\FrameworkBundle\Component\Domain\Config\DomainConfig;
 use Shopsys\FrameworkBundle\Component\Domain\Domain;
 use Shopsys\FrameworkBundle\Component\Plugin\PluginCrudExtensionFacade;
+use Shopsys\FrameworkBundle\Component\Transformers\IndexedBooleansToArrayOfIndexesTransformer;
 use Shopsys\FrameworkBundle\Form\DisplayOnlyType;
 use Shopsys\FrameworkBundle\Form\DomainsType;
 use Shopsys\FrameworkBundle\Form\FormRenderingConfigurationExtension;
@@ -149,6 +150,8 @@ class CategoryFormType extends AbstractType
                 'required' => false,
                 'label' => t('Display on'),
             ]);
+        $builderSettingsGroup->get('showOnDomains')
+            ->addModelTransformer(new IndexedBooleansToArrayOfIndexesTransformer());
 
         $builderSeoGroup = $builder->create('seo', GroupType::class, [
             'label' => t('Seo'),