+++ packages/framework/src/Form/Admin/Transport/TransportFormType.php
@@ -4,6 +4,7 @@ namespace Shopsys\FrameworkBundle\Form\Admin\Transport;
 
 use Ivory\CKEditorBundle\Form\Type\CKEditorType;
 use Shopsys\FormTypesBundle\YesNoType;
+use Shopsys\FrameworkBundle\Component\Transformers\IndexedBooleansToArrayOfIndexesTransformer;
 use Shopsys\FrameworkBundle\Form\DisplayOnlyType;
 use Shopsys\FrameworkBundle\Form\DomainsType;
 use Shopsys\FrameworkBundle\Form\GroupType;
@@ -97,6 +98,8 @@ class TransportFormType extends AbstractType
                 'empty_message' => t('You have to create some payment first.'),
                 'label' => t('Available payment methods'),
             ]);
+        $builderBasicInformationGroup->get('domains')
+            ->addModelTransformer(new IndexedBooleansToArrayOfIndexesTransformer());
 
         $builderAdditionalInformationGroup = $builder->create('additionalInformation', GroupType::class, [
             'label' => t('Additional information'),